Implement service in the controller
Define an implementation of the service and inject this service
implementation into the controller. In the controller, get items from
the service and pass it forward to the view.
In mustache template define an element that will represent an item.
Define elements with item's price and item's title. The test passes now.

+++ src/main/java/com/cakefactory/model/Item.java
@@ -1,4 +1,6 @@
 package com.cakefactory.model;
 
-public record Item(String title, double price) {
+import java.math.BigDecimal;
+
+public record Item(String title, BigDecimal price) {
 }
